'user' * update, fixes

+++ protected/controllers/SpecialityController.php
@@ -89,7 +89,7 @@ class SpecialityController extends Controller
                 'status' => 'failure',
                 'div' => $this->renderPartial('_form', array(
                     'model' => $model
-                ), true)
+                ), true, true)
             ));
             Yii::app()->end();
         }
@@ -134,7 +134,17 @@ class SpecialityController extends Controller
     {
         if (Yii::app()->request->isPostRequest)
         {
-            $this->loadModel($id)->delete();
+            if ($id === 'many')
+            {
+                if (isset($_POST['ids']) && is_array($_POST['ids']))
+                {
+                    foreach ($_POST['ids'] as $id)
+                        $this->loadModel($id)->delete();
+                }
+                Yii::app()->end();
+            }
+            else
+                $this->loadModel($id)->delete();
 
             if (!isset($_GET['ajax']))
                 $this->redirect($this->createUrl('index'));
